Add table header
