+++ views/sections/Watchlist/index.js
@@ -6,6 +6,7 @@ import { connect } from 'react-redux'
 import { toggleWatchlist } from 'ducks/watchlist/actions'
 import Header from 'components/Header'
 import { Eye } from 'components/icons/Eye'
+import TableHeader from 'components/TableHeader'
 import CoinCard from 'components/CoinCard'
 import Button from 'components/Button'
 import Footer from 'components/Footer'
@@ -81,33 +82,7 @@ const Watchlist = ({
               </Button>
             </div>
             <div className="home__table">
-              <div className="home__table-header aic jcsb">
-                <div className="table-item table-item_lg">
-                  <span className="p3 home__text home__text_number fw-medium">
-                    #
-                  </span>
-                  <span className="p3 fw-medium home__text">Coin name</span>
-                </div>
-                <span className="table-item fw-medium p3 home__text">
-                  Market cap
-                </span>
-                <span className="table-item table-item_sm fw-medium p3 home__text">
-                  Price
-                </span>
-                <span className="table-item fw-medium p3 home__text">
-                  Volume (24h)
-                </span>
-                <span className="table-item fw-medium p3 home__text">
-                  Circulating supply
-                </span>
-                <span className="table-item table-item_sm fw-medium p3 home__text">
-                  Change (24h)
-                </span>
-                <span className="table-item fw-medium p3 home__text home__text_ta_center">
-                  Price graph (7d)
-                </span>
-                <span className="home__empty" />
-              </div>
+              <TableHeader />
               {watchlist.data.map((crypto) => (
                 <CoinCard
                   order={crypto.cmc_rank}
@@ -153,33 +128,7 @@ const Watchlist = ({
             </div>
             <h2 className="h2 home__caption">Recommended coins:</h2>
             <div className="home__table">
-              <div className="home__table-header aic jcsb">
-                <div className="table-item table-item_lg">
-                  <span className="p3 home__text home__text_number fw-medium">
-                    #
-                  </span>
-                  <span className="p3 fw-medium home__text">Coin name</span>
-                </div>
-                <span className="table-item fw-medium p3 home__text">
-                  Market cap
-                </span>
-                <span className="table-item table-item_sm fw-medium p3 home__text">
-                  Price
-                </span>
-                <span className="table-item fw-medium p3 home__text">
-                  Volume (24h)
-                </span>
-                <span className="table-item fw-medium p3 home__text">
-                  Circulating supply
-                </span>
-                <span className="table-item table-item_sm fw-medium p3 home__text">
-                  Change (24h)
-                </span>
-                <span className="table-item fw-medium p3 home__text home__text_ta_center">
-                  Price graph (7d)
-                </span>
-                <span className="home__empty" />
-              </div>
+              <TableHeader />
               {watchlist.recommended.map((crypto) => (
                 <CoinCard
                   order={crypto.cmc_rank}
